+++ DarkStatsCore/Pages/Index.cshtml
@@ -65,14 +65,7 @@
 </div>
 <script src="~/lib/chartist/chartist.min.js"></script>
 <script src="~/lib/chartist/chartist-plugin-tooltip.min.js"></script>
-<script type="text/javascript">
-        if (typeof Promise === 'undefined') {
-            document.write('<script type="text/javascript" src="/lib/signalr/signalr-clientES5-1.0.0-alpha2-final.min.js"><\/script>');
-        }
-        else {
-            document.write('<script type="text/javascript" src="/lib/signalr/signalr-client-1.0.0-alpha2-final.min.js"><\/script>');
-        }
-    </script>
+<script src="~/lib/signalr/signalr.min.js"></script>
 <script>
 	var data = {
       series: [
@@ -97,8 +90,9 @@
     };
 
     var chart = new Chartist.Line(".ct-chart", data, options);
-    let logger = new signalR.ConsoleLogger(signalR.LogLevel.Information);
-    let connection = new signalR.HubConnection("/dashboard", { logger: logger });
+    let connection = new signalR.HubConnectionBuilder()
+        .withUrl("/dashboard")
+        .build();
     connection.start().then(function () {
 	    connection.invoke("GetDashboard").then(function (dashboard) {
             processDashboardStream(dashboard);
